requirements.txt
discord.py==2.2.3
selenium==4.9.1
